ABC142

+++ library/CPP/Primary.cpp
@@ -1,17 +1,17 @@
 // all input is natural num
 
 // primary judge O(sqrt(n))
-bool is_prime(int n) {
-    for (int i = 2; i * i <= n; i++) {
+bool is_prime(long long n) {
+    for (long long i = 2; i * i <= n; i++) {
         if (n % i == 0) return false;
     }
     return n != 1; // exclude 1
 }
 
 // make vector of divisor O(sqrt(n))
-vector<int> divisor(int n) {
-    vector<int> res;
-    for (int i = 1; i * i <= n; i++) {
+vector<long long> divisor(long long n) {
+    vector<long long> res;
+    for (long long i = 1; i * i <= n; i++) {
         if (n % i == 0) {
             res.push_back(i);
             if (i != n / i) res.push_back(n / i);
@@ -21,9 +21,9 @@ vector<int> divisor(int n) {
 }
 
 // factorization O(sqrt(n))
-map<int, int> prime_factor(int n) {
-    map<int, int> res;
-    for (int i = 2; i * i <= n; i++) {
+map<long long, long long> prime_factor(long long n) {
+    map<long long, long long> res;
+    for (long long i = 2; i * i <= n; i++) {
         while (n % i == 0) {
             ++res[i];
             n /= i;
@@ -34,19 +34,19 @@ map<int, int> prime_factor(int n) {
 }
 
 // Eratosthenes O(nloglog(n))
-int prime[MAX_N];     // i番目の素数
-bool is_prime[MAX_N]; // i が素数か否か　
+long long prime[MAX_N]; // i番目の素数
+bool is_prime[MAX_N];   // i が素数か否か　
 // n以下の素数の数を返す
-int sieve(int n) {
-    int p = 0;
-    for (int i = 0; i <= n; i++)
+long long sieve(long long n) {
+    long long p = 0;
+    for (long long i = 0; i <= n; i++)
         is_prime[i] = true;
     is_prime[0] = is_prime[1] = false;
-    for (int i = 2; i <= n; i++) {
+    for (long long i = 2; i <= n; i++) {
         if (is_prime[i]) {
             prime[p++] = i;
             // i の倍数全消去
-            for (int j = 2 * i; j <= n; j += i)
+            for (long long j = 2 * i; j <= n; j += i)
                 is_prime[j] = false;
         }
     }
@@ -54,9 +54,9 @@ int sieve(int n) {
 }
 
 struct Sieve {
-    int n;
-    vector<int> f, primes;
-    Sieve(int n) : n(n), f(n + 1) {
+    long long n;
+    vector<long long> f, primes;
+    Sieve(long long n) : n(n), f(n + 1) {
         f[0] = f[1] = -1;
         for (long long i = 2; i <= n; i++) {
             if (f[i]) continue;
@@ -67,20 +67,20 @@ struct Sieve {
             }
         }
     }
-    bool isPrime(int x) { return f[x] == x; }
-    vector<int> factorList(int x) {
-        vector<int> res;
+    bool isPrime(long long x) { return f[x] == x; }
+    vector<long long> factorList(long long x) {
+        vector<long long> res;
         while (x != 1) {
             res.push_back(f[x]);
             x /= f[x];
         }
         return res;
     }
-    vector<pair<int, int>> factor(int x) {
-        vector<int> fl = factorList(x);
+    vector<pair<long long, long long>> factor(long long x) {
+        vector<long long> fl = factorList(x);
         if (fl.size() == 0) return {};
-        vector<pair<int, int>> res(1, pair<int, int>(fl[0], 0));
-        for (int p : fl) {
+        vector<pair<long long, long long>> res(1, pair<long long, long long>(fl[0], 0));
+        for (long long p : fl) {
             if (res.back().first == p) {
                 res.back().second++;
             } else {
